Bug 855264 part 2 - Fix some MSVC warnings. r=djvj
UltraBlame original commit: 7d43d5cc057248d96aeb5826b88c855c0b7475f2

+++ js/src/ion/BaselineFrameInfo.cpp
@@ -72,7 +72,7 @@ FrameInfo::numUnsyncedSlots()
     
     uint32_t i = 0;
     for (; i < stackDepth(); i++) {
-        if (peek(-(i + 1))->kind() == StackValue::Stack)
+        if (peek(-int32_t(i + 1))->kind() == StackValue::Stack)
             break;
     }
     return i;
